refactor: getColorByImage pure function called in useSetImageAndNameFromPlatformData
