+++ src/components/create-guild/PickRolePlatform/hooks/useSetImageAndNameFromPlatformData.ts
@@ -1,3 +1,5 @@
+import Color from "color"
+import ColorThief from "colorthief/dist/color-thief.mjs"
 import { Dispatch, SetStateAction, useEffect } from "react"
 import { useFormContext, useFormState } from "react-hook-form"
 import getRandomInt from "utils/getRandomInt"
@@ -24,6 +26,8 @@ const useSetImageAndNameFromPlatformData = (
     }
 
     setValue("imageUrl", platformImage)
+    ;(async () => setValue("theme.color", await getColorByImage(platformImage)))()
+
     setUploadPromise(
       fetch(platformImage)
         .then((response) => response.blob())
@@ -44,5 +48,22 @@ const useSetImageAndNameFromPlatformData = (
     )
   }, [platformImage])
 }
+
+const getColorByImage = (imageUrl) =>
+  new Promise((resolve, reject) => {
+    const colorThief = new ColorThief()
+
+    const imgEl = document.createElement("img")
+    imgEl.src = imageUrl
+    imgEl.width = 64
+    imgEl.height = 64
+    imgEl.crossOrigin = "anonymous"
+
+    imgEl.addEventListener("load", () => {
+      const dominantRgbColor = colorThief.getColor(imgEl)
+      const dominantHexColor = Color.rgb(dominantRgbColor).hex()
+      resolve(dominantHexColor)
+    })
+  })
 
 export default useSetImageAndNameFromPlatformData
